Update util.py
Using os.path to save file to system as well as pickle.

Added auto saving that occurs every 60 seconds, as of now it does not detect if any changes were made and just changes. This is inefficient and will eventually be replaced with a smart saving system.

Added Font Family and Font Size variables in the Counter Display

Moved counter name validation to bottom of the CounterManager Class.

Renamed counters_dict to counter_dict